tests/requirements.txt
pytest==8.3.3
requests==2.32.3
mysql-connector-python==9.0.0
